chore(docs): Improved memory formatting in process monitoring UI
# Project State

## Current Focus
Improved memory formatting in process monitoring UI

## Context
The change addresses inconsistent memory display formatting by ensuring all memory values are treated as floating-point numbers for consistent formatting.

## Completed
- [x] Fixed memory display formatting to use `p.mem as f64` for consistent output
- [x] Updated Cargo.lock with dependency changes

## In Progress
- [ ] No active work in progress

## Blockers
- None identified

## Next Steps
1. Verify memory display consistency across different process states
2. Review if additional process metrics need similar formatting adjustments
---
category: chore
scope: docs
---

+++ src/ui/mod.rs
@@ -1695,7 +1695,7 @@ fn draw_processes_view(f: &mut Frame, area: Rect, app: &mut App) {
             })),
             Cell::from(p.status.clone()).style(Style::default().fg(status_col)),
             Cell::from(format!("{:.1}% {}", p.cpu, mini_bar(p.cpu as f64, 4))).style(Style::default().fg(cpu_color)),
-            Cell::from(format!("{} {}", format_memory_mib(p.mem), mem_bar)).style(Style::default().fg(mem_color)),
+            Cell::from(format!("{} {}", format_memory_mib(p.mem as f64), mem_bar)).style(Style::default().fg(mem_color)),
         ])
         .style(style)
     });
